Ensure PascalCase serialization for policy uploader (#12)
*Description of changes:* Adds PascalCase serialization for
`UploadResponse` and `BatchUploadResponse` which were missing it.
Extended a test to check the serialization of these types.

Note: We are making updates to ensure PascalCase serialization across
our public types, this PR covers the policy tools package.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.