Add service hints feature (#20)
*Issue #, if available:*

*Description of changes:*
* Added `--service-hints` feature with service synonym matching and "did
you mean" suggestion (based on edit-distance <= 2)
* Refactored service model index building (no need to call
`discover_services` anymore) and added caching to it
* Added service renaming for Smithy model names to Botocore model names
used for building the service synonyms: In Smithy it is e.g.
`cloudwatch-logs`, but in botocore it is `logs` (etc.). Furthermore, the
import in Go is `cloudwatchlogs`, in JS/TS `cloudwatch-logs`. All of
these are valid service-hints. I created the rename map by analyzing the
`serviceId`/`sdkId` in the model files. We should eventually add a
script which checks that this map is still up-to-date during CI.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.