refactor: SDK model loading (#55)
*Issue #, if available:*

*Description of changes:* Some cleanup in data loading: Moving to a
setup, where the `...Raw` structs have the raw embedded data, and the
`...Data` structs provide views on top of raw data. I think this
separation makes sense, as we sometimes combine data from multiple
files, and consumers of the data should not need to know which file it
comes from.
This also gets us a step closer to ensuring we load data into memory
only once.

Also removed `WaitersRegistry` and used `serde_json` directly instead of
the `JsonProvider`. If we decide to add a Wasm compilation target, we
will need to revisit this.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.